+++ test/projects/standard/src/store/create.js
@@ -1,21 +1,7 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux'
-import { reduxForCreateStore } from 'koot'
-import reducers from './reducers'
+import { createStore } from 'koot';
+import reducers from './reducers';
 
 /**
  * 项目自创建 store 的方法函数
  */
-export default () => {
-    const {
-        reducers: defaultReducers, initialState, middlewares
-    } = reduxForCreateStore
-
-    return createStore(
-        combineReducers({
-            ...defaultReducers,
-            ...reducers
-        }),
-        initialState,
-        applyMiddleware(...middlewares)
-    )
-}
+export default () => createStore(reducers);